fix(login): give feedback when trying to login with a Discord account that has no email bound to it
Users owning a Discord account which didn't have an email linked to that account, couldn't login to
the website. When they tried to login it would redirect to the home page without giving any feedback
to the user. Fix now outputs an error message prompt on entering the home page after a failed login
attempt.