fix(discord): fix error as access token cant be deleted before software instances (foreign key)
Database was throwing an error due in the attempt of deleting an access token when software
instances bound to that access token existed. This is because you can't delete a row where a value
is being referenced in another table as the foreign key. Error was solved by replacing delete access
token function with delete software instances function.